Fix missing link dependencies for X libraries.
Most of dependencies for X libraries were described as "build" and not "link".
However, I think this is wrong as when linking to this libraries, one has to be
linked also to compatible dependency. In Spack spirit, we have to use coherent
dependencies, that is the same. (In CMake packaging, however, the opposite
approach is advised: let the user choose the implementation of their choice,
without a priori tests.)